Added an EditorContext and ability to toggle markdown
The problem with disabling (obscuring markdown from the DOM) is that we cannot parse from the DOM anymore. WYSIWYG editors probably don’t have this problem because there is no associative syntax.

Actually, if we encode the syntax type on the element itself, the representation doesn’t matter anymore. In the case that we don’t know the syntax, we can fallback to the default/canonical syntax.