Implemented correctedPos pattern. Logic is slightly changed.
Instead of:

if (state.history.index === 1 && state.history.correctedPos) {
	state.history.correctedPos = false
}

Logic has been changed to:

if (state.history.index <= 1 && state.history.correctedPos) {
	state.history.correctedPos = false
}

The difference is that the editor corrects the pos on the first or second-to-first undo now. The reason for this is so idle timeout and storeUndo does not play a role with correcting pos.