Added temporary solution to prevent uselessly rerendering text nodes (prevents spellcheck from rerendering). This only works React attempts to diff an element with the same key.
Of course, if we are rendering a *different* key, React will blow it out of the DOM. So this does help with preventing spellcheck from rerendering only when keys are the same (e.g. we don’t increment the version number or timestamp). So this basically solves half the problem.

Now we need to communicate to React *when* we intend to render (or delete) React elements. This is the last step needed to make the editor React-friendly.