Save changes in anticipation of updating pos to support a root-document element offset:
const pos = {
	x: 0,    // X-axis offset (relative to y-axis offset)
	y: 0,    // Y-axis offset
	root: 0, // Root document-element offset
	pos: 0,  // Document position
}

This is needed for the native rendering strategy. The reason it’s preferred to compare root-level elements, as opposed to node-level elements, is to cover edge cases such as multiline strings in <Preformatted> elements where introducing a backtick, for example, requires a rerender. It’s not obvious that comparing elements on a node basis is sufficient or safe -- we know that comparing root elements is safe because that’s the behavior we’re mocking now.

There may be another way to achieve the same affect (albeit with more complexity). That is, to store data in nodes where a given node *knows* what type it belongs to.