Changed parseInlineElements to always return an array. This is important because it makes it easier to compare inside of areEqualInlineElements. Note that <Break> elements are exempt because they return children: null, which is fine because we handle that case already.
Now we have a somewhat naive but decent native rendering strategy **that only gets trigged for input and onCompositionEvents** because of renderOptions.id. Ideally we want to query the actual event type, like state.eventType === evenTypesEnum.

To create a less naive implementation, we would actually recurse to the current node decrementing x and look at how close we are to the next and previous inline elements. This is a reasonable approach.

The other thing is we were actually wrong about the multiline string inside of a <Preformatted> element, because we’re using dangerouslySetInnerHTML, which overrides React.