Found a pretty serious bug
It looks like lists, particularly checklists for example -- where the bug was first noticed -- can technically share references to old nodes because of mutateAnyListIDs works. This is still a best guess, but basically we’re only ever mutating an ID and not actually deeply copying anything.

The bug was observed with exactly the following:

```
- [ ] asdasdasdasd

- [x] asdasdasdasd

- [x] asdasdas das dasd asd asd asd asdasd
```

Pressing the first checkmark toggles the second checkmark instead.